fix json depth handlind

+++ app/config/tb_deploy_tool_controller.php
@@ -94,14 +94,15 @@ class Controller {
         if (file_exists('/var/www/deploy-tool/app/since_last_log.json') && !$dryRun) {
             $last_commits = json_decode(file_get_contents('/var/www/deploy-tool/app/since_last_log.json'), true);
 
-            if (end($last_commits['lastcommit']) != $last_log) {
-                $last_commits['lastcommit'][] = $last_log;
+            if (end($last_commits['lastcommit'][]) != $last_log) {
+                $last_commits['lastcommit'][][] = $last_log;
+                $last_commits['lastcommit'][][] = date('d/m/Y h:i');
             }
 
             file_put_contents('/var/www/deploy-tool/app/since_last_log.json', json_encode($last_commits));
 
         }
-        elseif (!file_exists('/var/www/deploy-tool/app/since_last_log.json')) {
+        elseif (!file_exists('/var/www/deploy-tool/app/since_last_log.json') && !$dryRun) {
             $lastcommit = array("lastcommit"=>array(array($last_log, date('d/m/Y h:i'))));
             file_put_contents('/var/www/deploy-tool/app/since_last_log.json', json_encode($lastcommit));
         }
